% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blast_Oligo_Candidates_Func.r
\name{blast_candidate_oligos}
\alias{blast_candidate_oligos}
\title{Blast Candidate Oligomers}
\usage{
blast_candidate_oligos(blastn_path = "blastn", blastdb, candidates,
  outfile = "Candidate_vs_UNITE.out")
}
\arguments{
\item{blastn_path}{Path to blastn. Defaults to "blastn". If your blast copy is somewhere else, please point to it here.}

\item{blastdb}{Filepath for the blast databse. No Default}

\item{candidates}{Filepath for the candidate oligomers file. No Default.}

\item{outfile}{Filepath for the output file name, which will contain the blast results. Defaults to "Candidates_vs_UNITE.out"}
}
\description{
This function will blast candidate oligomers (from a nontarget organism) against a database of target organisms and returns a parsable format.
}
\examples{
blast_candidate_oligos(blastdb="./Unite_s_011217/UNITE_s_011217")
}
\keyword{length}
\keyword{nontarget}
\keyword{outfile}
