% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Output_Func.r
\name{check_candidate_oligos}
\alias{check_candidate_oligos}
\title{Check Oligomers and Potential Primer Pairs}
\usage{
check_candidate_oligos(candidates, blastoutput,
  outputgraph = "Oligos_Plot.pdf", outputfile = "OligoPairs_Table.txt")
}
\arguments{
\item{candidates}{Filepath for the candidate oligomers file. No Default.}

\item{blastoutput}{Filepath for the output from blasting the candidates against the target database. No Default.}

\item{outputgraph}{Filepath for the output graph, which will show how many times each oligo hits target sequences. Defaults to "Oligos_Plot.pdf".}

\item{outputfile}{Filepath for the output table, which will check pairs of oligos and sort them by the minimum total number of times they hit target sequences. Defaults to "OligoPairs_Table.txt".}
}
\description{
This function takes output from get_candidate_oligos() and blast_candidate_oligos(). It checks each candidate for how many times it hit the target organisms, then sorts potential pairs to minimize the number of hits. The pairs in the output table can then be manually checked for ideal primer characteristics, etc.
}
\examples{
check_candidate_oligos(candidates="Candidate_oligos.fasta", blastoutput="Candidate_vs_UNITE.out")
}
\keyword{blastoutput}
\keyword{candidates}
\keyword{outputfile}
\keyword{outputgraph}
